Разделение на 3 модели: event-points, destinations, offers
