+++ src/model/destinations-model.js
@@ -1,0 +1,14 @@
+import { arrayDestinations } from '../mock/destinations.js';
+
+export default class DestinationsModel {
+  destinations = arrayDestinations;
+
+  getDestinations() {
+    return this.destinations;
+  }
+
+  getDestinationsById(id) {
+    const availableDestinations = this.getDestinations();
+    return availableDestinations.find((item) => item.id === id);
+  }
+}
