+++ src/presenter/main-presenter.js
@@ -10,21 +10,25 @@ import { render, RenderPosition } from '../render.js';
 export default class MainPresenter {
   eventList = new EventListView();
 
-  constructor({ container, eventPointModel }) {
+  constructor({ container, eventPointModel, offersModel, destinationsModel }) {
     this.container = container;
     this.eventPointModel = eventPointModel;
+    this.offersModel = offersModel;
+    this.destinationsModel = destinationsModel;
   }
 
   init() {
     this.eventsListPoints = [...this.eventPointModel.getEventPoint()];
+    this.offers = [...this.offersModel.getOffers()];
+    this.destinations = [...this.destinationsModel.getDestinations()];
 
     const editEventPoint = getRandomArrayElement(this.eventsListPoints);
     const formEditView = new FormEditView({
       eventPoint: editEventPoint,
-      availableOffers: this.eventPointModel.getOffersByType(editEventPoint.type),
-      pointDestination: this.eventPointModel.getDestinationsById(editEventPoint.destination),
-      destination: this.eventPointModel.getDestinations(),
-      arrayTypeOffers: this.eventPointModel.getOffersType(),
+      availableOffers: this.offersModel.getOffersByType(editEventPoint.type),
+      pointDestination: this.destinationsModel.getDestinationsById(editEventPoint.destination),
+      destination: this.destinationsModel.getDestinations(),
+      arrayTypeOffers: this.offersModel.getOffersType(),
     });
 
     render(new SortView(), this.container, RenderPosition.BEFOREEND);
@@ -35,8 +39,8 @@ export default class MainPresenter {
     for (let i = 0; i < this.eventsListPoints.length; i++) {
       const eventPoint = new LocationPointView({
         eventPoint: this.eventsListPoints[i],
-        destination: this.eventPointModel.getDestinationsById(this.eventsListPoints[i].destination),
-        offers: [...this.eventPointModel.getOffersById(this.eventsListPoints[i].type, this.eventsListPoints[i].offers)],
+        destination: this.destinationsModel.getDestinationsById(this.eventsListPoints[i].destination),
+        offers: [...this.offersModel.getOffersById(this.eventsListPoints[i].type, this.eventsListPoints[i].offers)],
       });
       render(eventPoint, this.eventList.getElement());
     }
